notification/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0
pydantic[email]==2.4.2